requirements.txt
openpyxl>=3.1.2
python-docx>=1.1.0
Pillow>=10.2.0
PyPDF2>=3.0.1
lxml>=5.1.0

